Move RCRunner-like things to test/utils
