+++ test/e2e/load.go
@@ -28,8 +28,7 @@ import (
 	"time"
 
 	"k8s.io/kubernetes/pkg/api"
-	"k8s.io/kubernetes/pkg/client/clientset_generated/internalclientset"
-	unversionedcore "k8s.io/kubernetes/pkg/client/clientset_generated/internalclientset/typed/core/unversioned"
+	clientset "k8s.io/kubernetes/pkg/client/clientset_generated/internalclientset"
 	"k8s.io/kubernetes/pkg/client/restclient"
 	"k8s.io/kubernetes/pkg/client/transport"
 	client "k8s.io/kubernetes/pkg/client/unversioned"
@@ -37,6 +36,7 @@ import (
 	"k8s.io/kubernetes/pkg/util/intstr"
 	utilnet "k8s.io/kubernetes/pkg/util/net"
 	"k8s.io/kubernetes/test/e2e/framework"
+	testutils "k8s.io/kubernetes/test/utils"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
@@ -66,7 +66,7 @@ var _ = framework.KubeDescribe("Load capacity", func() {
 	var c *client.Client
 	var nodeCount int
 	var ns string
-	var configs []*framework.RCConfig
+	var configs []*testutils.RCConfig
 	var namespaces []*api.Namespace
 
 	// Gathers metrics before teardown
@@ -223,8 +223,8 @@ func createNamespaces(f *framework.Framework, nodeCount, podsPerNode int) []*api
 	return namespaces
 }
 
-func createClients(numberOfClients int) ([]*client.Client, error) {
-	clients := make([]*client.Client, numberOfClients)
+func createClients(numberOfClients int) ([]clientset.Interface, error) {
+	clients := make([]clientset.Interface, numberOfClients)
 	for i := 0; i < numberOfClients; i++ {
 		config, err := framework.LoadConfig()
 		Expect(err).NotTo(HaveOccurred())
@@ -260,7 +260,7 @@ func createClients(numberOfClients int) ([]*client.Client, error) {
 		// Transport field.
 		config.TLSClientConfig = restclient.TLSClientConfig{}
 
-		c, err := client.New(config)
+		c, err := clientset.NewForConfig(config)
 		if err != nil {
 			return nil, err
 		}
@@ -283,8 +283,8 @@ func computeRCCounts(total int) (int, int, int) {
 	return smallRCCount, mediumRCCount, bigRCCount
 }
 
-func generateRCConfigs(totalPods int, image string, command []string, nss []*api.Namespace) []*framework.RCConfig {
-	configs := make([]*framework.RCConfig, 0)
+func generateRCConfigs(totalPods int, image string, command []string, nss []*api.Namespace) []*testutils.RCConfig {
+	configs := make([]*testutils.RCConfig, 0)
 
 	smallRCCount, mediumRCCount, bigRCCount := computeRCCounts(totalPods)
 	configs = append(configs, generateRCConfigsForGroup(nss, smallRCGroupName, smallRCSize, smallRCCount, image, command)...)
@@ -298,18 +298,18 @@ func generateRCConfigs(totalPods int, image string, command []string, nss []*api
 	framework.ExpectNoError(err)
 
 	for i := 0; i < len(configs); i++ {
-		configs[i].Client = clients[i%len(clients)]
+		configs[i].ClientSet = clients[i%len(clients)]
 	}
 
 	return configs
 }
 
 func generateRCConfigsForGroup(
-	nss []*api.Namespace, groupName string, size, count int, image string, command []string) []*framework.RCConfig {
-	configs := make([]*framework.RCConfig, 0, count)
+	nss []*api.Namespace, groupName string, size, count int, image string, command []string) []*testutils.RCConfig {
+	configs := make([]*testutils.RCConfig, 0, count)
 	for i := 1; i <= count; i++ {
-		config := &framework.RCConfig{
-			Client:     nil, // this will be overwritten later
+		config := &testutils.RCConfig{
+			ClientSet:  nil, // this will be overwritten later
 			Name:       groupName + "-" + strconv.Itoa(i),
 			Namespace:  nss[i%len(nss)].Name,
 			Timeout:    10 * time.Minute,
@@ -318,13 +318,14 @@ func generateRCConfigsForGroup(
 			Replicas:   size,
 			CpuRequest: 10,       // 0.01 core
 			MemRequest: 26214400, // 25MB
+			LogFunc:    framework.Logf,
 		}
 		configs = append(configs, config)
 	}
 	return configs
 }
 
-func generateServicesForConfigs(configs []*framework.RCConfig) []*api.Service {
+func generateServicesForConfigs(configs []*testutils.RCConfig) []*api.Service {
 	services := make([]*api.Service, 0, len(configs))
 	for _, config := range configs {
 		serviceName := config.Name + "-svc"
@@ -351,7 +352,7 @@ func sleepUpTo(d time.Duration) {
 	time.Sleep(time.Duration(rand.Int63n(d.Nanoseconds())))
 }
 
-func createAllRC(configs []*framework.RCConfig, creatingTime time.Duration) {
+func createAllRC(configs []*testutils.RCConfig, creatingTime time.Duration) {
 	var wg sync.WaitGroup
 	wg.Add(len(configs))
 	for _, config := range configs {
@@ -360,15 +361,15 @@ func createAllRC(configs []*framework.RCConfig, creatingTime time.Duration) {
 	wg.Wait()
 }
 
-func createRC(wg *sync.WaitGroup, config *framework.RCConfig, creatingTime time.Duration) {
+func createRC(wg *sync.WaitGroup, config *testutils.RCConfig, creatingTime time.Duration) {
 	defer GinkgoRecover()
 	defer wg.Done()
 
 	sleepUpTo(creatingTime)
-	framework.ExpectNoError(framework.RunRC(*config), fmt.Sprintf("creating rc %s", config.Name))
+	framework.ExpectNoError(testutils.RunRC(*config), fmt.Sprintf("creating rc %s", config.Name))
 }
 
-func scaleAllRC(configs []*framework.RCConfig, scalingTime time.Duration) {
+func scaleAllRC(configs []*testutils.RCConfig, scalingTime time.Duration) {
 	var wg sync.WaitGroup
 	wg.Add(len(configs))
 	for _, config := range configs {
@@ -379,24 +380,24 @@ func scaleAllRC(configs []*framework.RCConfig, scalingTime time.Duration) {
 
 // Scales RC to a random size within [0.5*size, 1.5*size] and lists all the pods afterwards.
 // Scaling happens always based on original size, not the current size.
-func scaleRC(wg *sync.WaitGroup, config *framework.RCConfig, scalingTime time.Duration) {
+func scaleRC(wg *sync.WaitGroup, config *testutils.RCConfig, scalingTime time.Duration) {
 	defer GinkgoRecover()
 	defer wg.Done()
 
 	sleepUpTo(scalingTime)
 	newSize := uint(rand.Intn(config.Replicas) + config.Replicas/2)
-	framework.ExpectNoError(framework.ScaleRC(config.Client, coreClientSetFromUnversioned(config.Client), config.Namespace, config.Name, newSize, true),
+	framework.ExpectNoError(framework.ScaleRC(config.ClientSet, config.Namespace, config.Name, newSize, true),
 		fmt.Sprintf("scaling rc %s for the first time", config.Name))
 	selector := labels.SelectorFromSet(labels.Set(map[string]string{"name": config.Name}))
 	options := api.ListOptions{
 		LabelSelector:   selector,
 		ResourceVersion: "0",
 	}
-	_, err := config.Client.Pods(config.Namespace).List(options)
+	_, err := config.ClientSet.Core().Pods(config.Namespace).List(options)
 	framework.ExpectNoError(err, fmt.Sprintf("listing pods from rc %v", config.Name))
 }
 
-func deleteAllRC(configs []*framework.RCConfig, deletingTime time.Duration) {
+func deleteAllRC(configs []*testutils.RCConfig, deletingTime time.Duration) {
 	var wg sync.WaitGroup
 	wg.Add(len(configs))
 	for _, config := range configs {
@@ -405,25 +406,14 @@ func deleteAllRC(configs []*framework.RCConfig, deletingTime time.Duration) {
 	wg.Wait()
 }
 
-func deleteRC(wg *sync.WaitGroup, config *framework.RCConfig, deletingTime time.Duration) {
+func deleteRC(wg *sync.WaitGroup, config *testutils.RCConfig, deletingTime time.Duration) {
 	defer GinkgoRecover()
 	defer wg.Done()
 
 	sleepUpTo(deletingTime)
 	if framework.TestContext.GarbageCollectorEnabled {
-		framework.ExpectNoError(framework.DeleteRCAndWaitForGC(config.Client, config.Namespace, config.Name), fmt.Sprintf("deleting rc %s", config.Name))
+		framework.ExpectNoError(framework.DeleteRCAndWaitForGC(config.ClientSet, config.Namespace, config.Name), fmt.Sprintf("deleting rc %s", config.Name))
 	} else {
-		framework.ExpectNoError(framework.DeleteRCAndPods(config.Client, coreClientSetFromUnversioned(config.Client), config.Namespace, config.Name), fmt.Sprintf("deleting rc %s", config.Name))
+		framework.ExpectNoError(framework.DeleteRCAndPods(config.ClientSet, config.Namespace, config.Name), fmt.Sprintf("deleting rc %s", config.Name))
 	}
-}
-
-// coreClientSetFromUnversioned adapts just enough of a a unversioned.Client to work with the scale RC function
-func coreClientSetFromUnversioned(c *client.Client) internalclientset.Interface {
-	var clientset internalclientset.Clientset
-	if c != nil {
-		clientset.CoreClient = unversionedcore.New(c.RESTClient)
-	} else {
-		clientset.CoreClient = unversionedcore.New(nil)
-	}
-	return &clientset
 }
