+++ test/e2e/proxy.go
@@ -31,6 +31,7 @@ import (
 	"k8s.io/kubernetes/pkg/util/intstr"
 	"k8s.io/kubernetes/pkg/util/net"
 	"k8s.io/kubernetes/test/e2e/framework"
+	testutils "k8s.io/kubernetes/test/utils"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
@@ -116,8 +117,8 @@ func proxyContext(version string) {
 		// environmental variables below.
 		By("starting an echo server on multiple ports")
 		pods := []*api.Pod{}
-		cfg := framework.RCConfig{
-			Client:       f.Client,
+		cfg := testutils.RCConfig{
+			ClientSet:    f.ClientSet,
 			Image:        "gcr.io/google_containers/porter:cd5cb5791ebaa8641955f0e8c2a9bed669b1eaab",
 			Name:         service.Name,
 			Namespace:    f.Namespace.Name,
@@ -152,9 +153,10 @@ func proxyContext(version string) {
 			},
 			Labels:      labels,
 			CreatedPods: &pods,
+			LogFunc:     framework.Logf,
 		}
-		Expect(framework.RunRC(cfg)).NotTo(HaveOccurred())
-		defer framework.DeleteRCAndPods(f.Client, f.ClientSet, f.Namespace.Name, cfg.Name)
+		Expect(testutils.RunRC(cfg)).NotTo(HaveOccurred())
+		defer framework.DeleteRCAndPods(f.ClientSet, f.Namespace.Name, cfg.Name)
 
 		Expect(f.WaitForAnEndpoint(service.Name)).NotTo(HaveOccurred())
