+++ test/e2e/service_latency.go
@@ -29,6 +29,7 @@ import (
 	"k8s.io/kubernetes/pkg/util/sets"
 	"k8s.io/kubernetes/pkg/watch"
 	"k8s.io/kubernetes/test/e2e/framework"
+	testutils "k8s.io/kubernetes/test/utils"
 
 	. "github.com/onsi/ginkgo"
 )
@@ -115,18 +116,18 @@ var _ = framework.KubeDescribe("Service endpoints latency", func() {
 })
 
 func runServiceLatencies(f *framework.Framework, inParallel, total int) (output []time.Duration, err error) {
-	cfg := framework.RCConfig{
-		Client:       f.Client,
+	cfg := testutils.RCConfig{
+		ClientSet:    f.ClientSet,
 		Image:        framework.GetPauseImageName(f.Client),
 		Name:         "svc-latency-rc",
 		Namespace:    f.Namespace.Name,
 		Replicas:     1,
 		PollInterval: time.Second,
 	}
-	if err := framework.RunRC(cfg); err != nil {
+	if err := testutils.RunRC(cfg); err != nil {
 		return nil, err
 	}
-	defer framework.DeleteRCAndPods(f.Client, f.ClientSet, f.Namespace.Name, cfg.Name)
+	defer framework.DeleteRCAndPods(f.ClientSet, f.Namespace.Name, cfg.Name)
 
 	// Run a single watcher, to reduce the number of API calls we have to
 	// make; this is to minimize the timing error. It's how kube-proxy
